+++ packages/terra-form-field/src/Field.jsx
@@ -44,6 +44,7 @@ const propTypes = {
    */
   isInline: PropTypes.bool,
   /**
+   * @private
    * Whether or not the label is visible. Use this props to hide a label while still creating it on the DOM for accessibility.
    */
   isLabelHidden: PropTypes.bool,