+++ packages/terra-form-input/src/InputField.jsx
@@ -37,6 +37,7 @@ const propTypes = {
    */
   help: PropTypes.node,
   /**
+   * @private
    * Whether or not to hide the required indicator on the label.
    */
   hideRequired: PropTypes.bool,
@@ -58,6 +59,7 @@ const propTypes = {
    */
   isInvalid: PropTypes.bool,
   /**
+   * @private
    * Whether or not the label is visible. Use this props to hide a label while still creating it on the DOM for accessibility.
    */
   isLabelHidden: PropTypes.bool,