+++ packages/terra-form-input/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+* Changed
+  * Set the hideRequired prop to be private
+  * Set the isLabelHidden prop to be private
+
 ## 4.17.5 - (July 5, 2022)
 
 * Changed