+++ packages/terra-form-input/CHANGELOG.md
@@ -5,6 +5,7 @@
 * Changed
   * Set the hideRequired prop to be private
   * Set the isLabelHidden prop to be private
+  * Set the ariaLabel prop to be private
 
 ## 4.17.5 - (July 5, 2022)